requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
Jinja2==3.1.2
gunicorn==21.2.0
openai>=1.30.0
python-dotenv>=1.0.1
